# backend/openapi.yaml
openapi: 3.0.0
info:
  title: QQM API
  version: 1.0.0
  description: Client / project / infrastructure management API. Generated from
    the actual implemented routes, controllers, and zod validators
    (backend/src/openapi) — not from any prior design document.
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: /
    description: Same-origin backend
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT obtained from POST /api/auth/login, sent as `Authorization:
        Bearer <token>`."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - message
      required:
        - error
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required:
        - page
        - per_page
        - total
        - total_pages
...
    ServerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Server"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    Server:
      type: object
      properties:
        id:
          type: string
          format: uuid
        environment_id:
          type: string
          format: uuid
        hostname:
          type: string
        ip_address:
          type: string
          nullable: true
        tech_stack:
          type: array
          items:
            type: string
        monitoring_url:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          nullable: true
          format: date-time
        display_name:
          type: string
        service_type:
          $ref: "#/components/schemas/ServiceType"
        access_method:
          $ref: "#/components/schemas/AccessMethod"
        access_host:
          type: string
        access_port:
          type: integer
          nullable: true
        access_path:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      required:
        - id
        - environment_id
        - hostname
        - ip_address
        - tech_stack
        - monitoring_url
        - notes
        - created_at
        - updated_at
        - deleted_at
        - display_name
        - service_type
        - access_method
        - access_host
        - access_port
        - access_path
        - username
        - password
    ServiceType:
      type: string
      nullable: true
      enum:
        - application
        - database
        - proxy
        - monitoring
        - repository
        - metrics
        - jump_host
        - other
        - null
    AccessMethod:
      type: string
      nullable: true
      enum:
        - ssh
        - rdp
        - telnet
        - web
        - other
        - null
    ServerDetail:
      allOf:
        - $ref: "#/components/schemas/Server"
        - type: object
          properties:
            environment:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
                project:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    name:
                      type: string
                  required:
                    - id
                    - name
              required:
                - id
                - name
                - project
          required:
            - environment
...
paths:
...
  /api/servers:
    get:
      tags:
        - Servers
      operationId: listServers
      summary: List servers
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            description: "Default: 1"
          required: false
          description: "Default: 1"
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            description: "Default: 20, max: 100"
          required: false
          description: "Default: 20, max: 100"
          name: per_page
          in: query
        - schema:
            type: string
            enum:
              - display_name
              - created_at
              - updated_at
            description: "Default: display_name"
          required: false
          description: "Default: display_name"
          name: sort
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            description: "Default: asc"
          required: false
          description: "Default: asc"
          name: order
          in: query
        - schema:
            type: string
            enum:
              - "false"
              - "true"
              - all
            description: "Default: false"
          required: false
          description: "Default: false"
          name: deleted
          in: query
        - schema:
            type: string
          required: false
          name: search
          in: query
        - schema:
            type: string
            format: uuid
          required: false
          name: environment_id
          in: query
        - schema:
            type: string
          required: false
          name: service_type
          in: query
        - schema:
            type: string
          required: false
          name: access_method
          in: query
      responses:
        "200":
          description: Paginated servers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Servers
      operationId: createServer
      summary: Create a server (admin or member)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                environment_id:
                  type: string
                  format: uuid
                display_name:
                  type: string
                  minLength: 1
                hostname:
                  type: string
                  minLength: 1
                ip_address:
                  type: string
                service_type:
                  type: string
                  minLength: 1
                  maxLength: 100
                access_method:
                  type: string
                  enum:
                    - ssh
                    - rdp
                    - telnet
                    - web
                    - other
                access_host:
                  type: string
                  minLength: 1
                access_port:
                  type: integer
                  minimum: 1
                  maximum: 65535
                access_path:
                  type: string
                tech_stack:
                  type: array
                  items:
                    type: string
                  default: []
                monitoring_url:
                  type: string
                  format: uri
                username:
                  type: string
                password:
                  type: string
                notes:
                  type: string
              required:
                - environment_id
                - display_name
                - hostname
      responses:
        "201":
          description: Server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/servers/{id}:
    get:
      tags:
        - Servers
      operationId: getServerById
      summary: Get a server by id, with its environment and that environment's project
        inlined
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Servers
      operationId: updateServer
      summary: Update a server (admin or member; optimistic lock; environment_id is
        not re-parentable)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  minLength: 1
                hostname:
                  type: string
                  minLength: 1
                ip_address:
                  type: string
                service_type:
                  type: string
                  minLength: 1
                  maxLength: 100
                access_method:
                  type: string
                  enum:
                    - ssh
                    - rdp
                    - telnet
                    - web
                    - other
                access_host:
                  type: string
                  minLength: 1
                access_port:
                  type: integer
                  minimum: 1
                  maximum: 65535
                access_path:
                  type: string
                tech_stack:
                  type: array
                  items:
                    type: string
                monitoring_url:
                  type: string
                  format: uri
                username:
                  type: string
                password:
                  type: string
                notes:
                  type: string
                updated_at:
                  type: string
                  minLength: 1
              required:
                - updated_at
      responses:
        "200":
          description: Server updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Servers
      operationId: deleteServer
      summary: Soft-delete a server (admin only; hard-deletes its credential_references)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Server soft-deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/servers/{id}/restore:
    post:
      tags:
        - Servers
      operationId: restoreServer
      summary: Restore a soft-deleted server (admin only; credential_references
        hard-deleted at delete time are NOT restored)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Server restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...